+++ src/Client.php
@@ -21,7 +21,7 @@ use Pronamic\WordPress\Pay\Core\XML\Security;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.0.9
+ * @version 2.1.4
  * @since   1.0.0
  */
 class Client {
@@ -317,6 +317,25 @@ class Client {
 		);
 
 		return $response;
+	}
+
+	/**
+	 * Get mandate.
+	 *
+	 * @param string $mandate_id Mollie mandate ID.
+	 * @return object
+	 * @throws \InvalidArgumentException Throws exception on empty mandate ID argument.
+	 */
+	public function get_mandate( $mandate_id, $customer_id ) {
+		if ( '' === $mandate_id ) {
+			throw new \InvalidArgumentException( 'Mollie mandate ID can not be empty string.' );
+		}
+
+		if ( '' === $customer_id ) {
+			throw new \InvalidArgumentException( 'Mollie customer ID can not be empty string.' );
+		}
+
+		return $this->send_request_to_endpoint( 'customers/' . $customer_id . '/mandates/' . $mandate_id, 'GET' );
 	}
 
 	/**